fix: use bigint for pagination and version params to prevent int32 overflow (G115)
- Cast LIMIT/OFFSET to bigint in GetApplicationsPaged SQL query so sqlc
  generates int64 fields (Column1/Column2) instead of int32, eliminating
  G115 on the int->int32 conversions at the call site
- Change version_major/version_minor/version_patch schema columns from
  INTEGER to BIGINT so sqlc generates int64 fields in the Release model,
  eliminating G115 on the int64->int32 casts in modelToPgUpsertRelease
- Update ListReleasesPaged manual scan variables from int32 to int64 to
  match the new BIGINT column types
- Add a safe uint64->int64 helper in parseSemverParts (caps at MaxInt64)
  with #nosec G115 annotation to suppress the residual gosec finding on
  the bounded conversion; real semver numbers never approach MaxInt64
- Update postgres_test.go version sort column assertions from int32 to
  int64 to match the regenerated model types

+++ internal/storage/postgres_test.go
@@ -669,9 +669,9 @@ func TestPostgresStorageSaveRelease_VersionSortColumns(t *testing.T) {
 	tests := []struct {
 		name         string
 		version      string
-		wantMajor    int32
-		wantMinor    int32
-		wantPatch    int32
+		wantMajor    int64
+		wantMinor    int64
+		wantPatch    int64
 		wantPreValid bool
 		wantPreStr   string
 	}{
