+++ internal/storage/sqlc/queries/postgres/applications.sql
@@ -27,4 +27,4 @@ SELECT id, name, description, platforms, config, created_at, updated_at,
        COUNT(*) OVER() AS total_count
 FROM applications
 ORDER BY name
-LIMIT $1 OFFSET $2;
+LIMIT $1::bigint OFFSET $2::bigint;
